Initial commit
This is an OpenRS2 deob of revision 225 with a lot of work done to rename and clean up the output. I've matched a bunch to the original Jagex names and included a very simple server that implements the bare minimum to run around. The server shares code with the client.

Some improvements have been made as well; I fixed a transparency bug, added bilinear filtering, and started replacing a lot of magic numbers with constants.